requirements.txt
google-cloud-storage==1.42.2
requests==2.26.0
